move tile to list-and-tile
