+++ src/app/e2e-testee/combo-select/change/app.module.ts
@@ -1,6 +1,6 @@
 import {NgModule} from "@angular/core";
 import {ComboSelectChangeDemo} from "./app.component";
-import {JigsawTileSelectModule} from "jigsaw/component/tile/tile";
+import {JigsawTileSelectModule} from "jigsaw/component/list-and-tile/tile";
 import {JigsawComboSelectModule} from "jigsaw/component/combo-select";
 import {CommonModule} from "@angular/common";
 @NgModule({
