+++ src/app/e2e-testee/combo-select/disable/app.module.ts
@@ -1,5 +1,5 @@
 import {NgModule} from "@angular/core";
-import {JigsawTileSelectModule} from "jigsaw/component/tile/tile";
+import {JigsawTileSelectModule} from "jigsaw/component/list-and-tile/tile";
 import {JigsawButtonModule} from "jigsaw/component/button/button";
 import {JigsawComboSelectModule} from "jigsaw/component/combo-select/index";
 import {DisabledComboSelectDemo} from "./app.component";
