+++ src/app/e2e-testee/combo-select/setWidth/app.module.ts
@@ -1,5 +1,5 @@
 import {NgModule} from "@angular/core";
-import {JigsawTileSelectModule} from "jigsaw/component/tile/tile";
+import {JigsawTileSelectModule} from "jigsaw/component/list-and-tile/tile";
 import {JigsawInputModule} from "jigsaw/component/input/input";
 import {JigsawComboSelectModule} from "jigsaw/component/combo-select/index";
 import {ComboSelectSetWidthDemo} from "./app.component";
