+++ src/app/e2e-testee/range-time/limitStart/app.module.ts
@@ -1,6 +1,6 @@
 import {NgModule} from "@angular/core";
 import {JigsawRangeTimeModule} from "jigsaw/component/range-time/index";
-import {JigsawTileSelectModule} from "jigsaw/component/tile/tile";
+import {JigsawTileSelectModule} from "jigsaw/component/list-and-tile/tile";
 import {RangeTimeLimitStartComponent} from "./app.component";
 import {CommonModule} from "@angular/common";
 @NgModule({
