+++ src/app/e2e-testee/tileselect/multipleSelect/app.module.ts
@@ -1,5 +1,5 @@
 import { NgModule } from '@angular/core';
-import { JigsawTileSelectModule } from "jigsaw/component/tile/tile";
+import { JigsawTileSelectModule } from "jigsaw/component/list-and-tile/tile";
 import { TileselectMultipleSelectDemoComponent }  from './app.component';
 import {CommonModule} from "@angular/common";
