+++ src/app/e2e-testee/time/limitEnd/app.module.ts
@@ -1,7 +1,7 @@
 import { NgModule } from '@angular/core';
 import { CommonModule } from "@angular/common";
 import { JigsawTimeModule } from "jigsaw/component/time/index";
-import { JigsawTileSelectModule } from "jigsaw/component/tile/tile";
+import { JigsawTileSelectModule } from "jigsaw/component/list-and-tile/tile";
 import { TimeLimitEndComponent }  from './app.component';
 
 @NgModule({
