+++ src/app/live-demo/list/app.module.ts
@@ -1,6 +1,6 @@
 import {NgModule} from "@angular/core";
 import {ListBasicDemoComponent} from "./app.component";
-import {JigsawListModule} from "jigsaw/component/list/list";
+import {JigsawListModule} from "jigsaw/component/list-and-tile/list";
 import {CommonModule} from "@angular/common";
 import {JigsawCheckBoxModule} from "../../../jigsaw/component/checkbox/index";
