+++ src/app/live-demo/tile/app.module.ts
@@ -1,5 +1,5 @@
 import { NgModule } from '@angular/core';
-import { JigsawTileSelectModule } from "jigsaw/component/tile/tile";
+import { JigsawTileSelectModule } from "jigsaw/component/list-and-tile/tile";
 import { TileselectBasicDemoComponent }  from './app.component';
 import {CommonModule} from "@angular/common";
 import {JigsawSwitchModule} from "jigsaw/component/switch/index";
