+++ src/jigsaw/component/list-and-tile/list-option.scss
