+++ src/jigsaw/core/theming/all-theme.scss
@@ -28,12 +28,12 @@
 @import "../../component/table/table-head";
 @import "../../component/tabs/tab";
 @import "../../component/tag/tag";
-@import "../../component/tile/tile";
-@import "../../component/tile/tile-option";
+@import "../../component/list-and-tile/tile";
+@import "../../component/list-and-tile/tile-option";
 @import "../../component/time/time";
 @import "../../component/tooltip/tooltip";
-@import "../../component/list/list";
-@import "../../component/list/list-option";
+@import "../../component/list-and-tile/list";
+@import "../../component/list-and-tile/list-option";
